Created parametric model of the Cessaroni 24-54mm casings.
This model is valid from the 24mm CTI casing to the 54mm CTI casing. Does not include XL variants or rear closure geometry. The model is accurate to within +/- 10% of the actual casing mass. There is an Issue with the combination of 3 grains and 29mm which should get fixed soon.